allow to reuse threads in quarantine
We can die with out of threads when we have 512 threads in quarantine.
This is not right. Allow to reuse threads in quarantine.

Always allocate threads from pool first. In user-space TSAN we prefer
to reuse threads over allocating new because it reduces memory consumption.
In kernel we have all threads pre-allocated, so there is no point in
reusing when we have brand new threads.